refactor: create `model` folder
