+++ lib/tuner/tuner.dart
@@ -5,7 +5,7 @@ import 'dart:typed_data';
 
 import 'package:flutter_audio_capture/flutter_audio_capture.dart';
 import 'package:mic_stream/mic_stream.dart';
-import 'package:musbx/tuner/note.dart';
+import 'package:musbx/model/note.dart';
 import 'package:pitch_detector_dart/pitch_detector.dart';
 import 'package:pitch_detector_dart/pitch_detector_result.dart';
